Fix Autoboot-dialogue showing twice with nokexec
With nokexec workaround, when we boot to secondary we put the secondary bootimg
to the primary and boot. The second boot, will not consider this transient state
and we end up showing Autoboot-dialogue again.

Store the state if the new primary is in progress, and skip the ROM selection
the next reboot. In this way, the ROM selection is done once before we
completely boot into secondary ROM.