Fix, booting external ROMs from recovery went loop
From Recovery, if we reboot and during Autoboot-Dialogue External SDCard ROM is
selected, then it goes into never-ending Autoboot Loop.
The reason is, USB/SDCard ROMs are not mounted if the is_second_boot is not set,
and so everytime it tries to match the current ROM (SDCard) it can never match
with other Internal ROMs, and keeps Looping.

Not sure, if anything to do with Device specific issue, but, now the USB/SDCard
ROMs population is done without checking is_second_boot (only safe check now
then would be if the current rom being booted contains partition info).

It is bettern than Loop, or Twice Autoboot-dialogue or any other artifacts, than
taking a bit time to load the USB/SDCard ROMs.